:scissors: Remove code duplication
