+++ app/lib/utils/getBaseQuery.js
@@ -1,0 +1,14 @@
+function getBaseQuery() {
+  return {
+    index: 'profiles',
+    type: 'gps',
+    body: {
+      size: 30,
+      query: {
+        bool: {}
+      }
+    }
+  };
+}
+
+module.exports = getBaseQuery;
